+++ src/System.Runtime.Serialization.Formatters/tests/BinaryFormatterTests.cs
@@ -113,7 +113,7 @@ namespace System.Runtime.Serialization.Formatters.Tests
             }
 
             // We store our framework blobs in index 1
-            int platformBlobIndex = TypeSerializableValue.GetPlatformIndex(blobs);
+            int platformBlobIndex = blobs.GetPlatformIndex();
             for (int i = 0; i < blobs.Length; i++)
             {
                 // Check if the current blob is from the current running platform.
@@ -549,7 +549,7 @@ namespace System.Runtime.Serialization.Formatters.Tests
             }
 
             // Check if runtime generated blob is the same as the stored one
-            int frameworkBlobNumber = TypeSerializableValue.GetPlatformIndex(blobs);
+            int frameworkBlobNumber = blobs.GetPlatformIndex();
             if (frameworkBlobNumber < blobs.Length)
             {
                 string runtimeBlob = BinaryFormatterHelpers.ToBase64String(obj, FormatterAssemblyStyle.Full);
