+++ src/System.Runtime.Serialization.Formatters/tests/System.Runtime.Serialization.Formatters.Tests.csproj
@@ -19,6 +19,7 @@
     <Compile Include="FormatterServicesTests.cs" />
     <Compile Include="FormatterServicesTests.Windows.cs" Condition="'$(TargetsWindows)' == 'true'" />
     <Compile Include="FormatterTests.cs" />
+    <Compile Include="PlatformExtensions.cs" />
     <Compile Include="SerializationBinderTests.cs" />
     <Compile Include="SerializationInfoTests.cs" />
     <Compile Include="SerializationTypes.cs" />
