+++ src/System.Runtime.Serialization.Formatters/tests/TypeSerializableValue.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
-using System.Linq;
 
 namespace System.Runtime.Serialization.Formatters.Tests
 {
@@ -18,53 +16,6 @@ namespace System.Runtime.Serialization.Formatters.Tests
         {
             Base64Blob = base64Blob;
             Platform = platform;
-        }
-
-        public static int GetPlatformIndex(TypeSerializableValue[] blobs)
-        {
-            List<TypeSerializableValue> blobList = blobs.ToList();
-
-            // .NET Framework
-            if (PlatformDetection.IsFullFramework)
-            {
-                // Check if a specialized blob for >=netfx472 is present and return if found.
-                if (PlatformDetection.IsNetfx472OrNewer)
-                {
-                    int index = blobList.FindIndex(b => b.Platform == TargetFrameworkMoniker.netfx472);
-
-                    if (index >= 0)
-                        return index;
-                }
-
-                // Check if a specialized blob for >=netfx471 is present and return if found.
-                if (PlatformDetection.IsNetfx471OrNewer)
-                {
-                    int index = blobList.FindIndex(b => b.Platform == TargetFrameworkMoniker.netfx471);
-
-                    if (index >= 0)
-                        return index;
-                }
-                    
-                // If no newer blob for >=netfx471 is present use existing one. 
-                // If no netfx blob is present then -1 will be returned.
-                return blobList.FindIndex(b => b.Platform == TargetFrameworkMoniker.netfx461);
-            }
-
-            // .NET Core
-            if (PlatformDetection.IsNetCore)
-            {
-                // Check if a specialized blob for >=netcoreapp2.1 is present and return if found.
-                int index = blobList.FindIndex(b => b.Platform == TargetFrameworkMoniker.netcoreapp21);
-
-                if (index >= 0)
-                    return index;
-
-                // If no newer blob for >=netcoreapp2.1 is present use existing one.
-                // If no netfx blob is present then -1 will be returned.
-                return blobList.FindIndex((b => b.Platform == TargetFrameworkMoniker.netcoreapp20));
-            }
-
-            return -1;
         }
     }
 }
